pyproject.toml
[project]
name = "jobconnect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "grandalf>=0.8",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.2.14",
    "langgraph>=0.6.6",
    "lxml>=6.0.1",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
]
